✨ Splash 화면 애니메이션 개선 및 기본 UI 수정 🔧 비밀번호 재설정 화면 팝업 로직 추가, API 호출 주석 처리

+++ data/src/main/java/com/example/data/implementation/repository/UserRepositoryImpl.kt
@@ -53,7 +53,7 @@ class UserRepositoryImpl @Inject constructor(
             email = email,
             password = password,
             gender = gender,
-            jobId = jobId.toLong(),
+            jobId = jobId,
             purposeList = purposeList,
             interestList = interestList
         )
